Move CZ to controlled_ops
