+++ pennylane/ops/op_math/__init__.py
@@ -75,7 +75,7 @@ Decompositions
 from .adjoint import Adjoint, adjoint
 from .composite import CompositeOp
 from .controlled import Controlled, ControlledOp, ctrl
-from .controlled_ops import ControlledQubitUnitary
+from .controlled_ops import ControlledQubitUnitary, CZ
 from .evolution import Evolution
 from .exp import Exp, exp
 from .pow import Pow, pow
@@ -85,6 +85,4 @@ from .sum import Sum, sum
 from .symbolicop import ScalarSymbolicOp, SymbolicOp
 from .controlled_decompositions import ctrl_decomp_zyz, ctrl_decomp_bisect
 
-controlled_qubit_ops = {
-    "ControlledQubitUnitary",
-}
+controlled_qubit_ops = {"ControlledQubitUnitary", "CZ"}
