+++ pennylane/ops/qubit/__init__.py
@@ -57,7 +57,6 @@ ops = {
     "T",
     "SX",
     "CNOT",
-    "CZ",
     "CCZ",
     "CY",
     "CH",
